Rename Potterhat.Node to PotterhatNode
