+++ apps/potterhat_node/README.md
@@ -1,3 +1,3 @@
-# Potterhat.Node
+# PotterhatNode
 
 Consumes Ethereum event and data from Ethereum nodes.
