+++ apps/potterhat_node/lib/node/node_config.ex
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-defmodule Potterhat.Node.NodeConfig do
+defmodule PotterhatNode.NodeConfig do
   @enforce_keys [:id, :label, :client, :rpc, :ws, :priority]
   defstruct [:id, :label, :client, :rpc, :ws, :priority]
