+++ apps/potterhat_node/lib/node/subscription/log.ex
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-defmodule Potterhat.Node.Subscription.Log do
+defmodule PotterhatNode.Subscription.Log do
   use WebSockex
 
   @subscription_id 3
