+++ apps/potterhat_node/lib/node/subscription/new_head.ex
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-defmodule Potterhat.Node.Subscription.NewHead do
+defmodule PotterhatNode.Subscription.NewHead do
   use WebSockex
 
   @subscription_id 1
