+++ apps/potterhat_node/lib/node/subscription/new_pending_transaction.ex
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-defmodule Potterhat.Node.Subscription.NewPendingTransaction do
+defmodule PotterhatNode.Subscription.NewPendingTransaction do
   use WebSockex
 
   @subscription_id 4
