+++ apps/potterhat_node/test/support/mock_ethereum_node/web_socket.ex
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-defmodule Potterhat.Node.MockEthereumNode.WebSocket do
+defmodule PotterhatNode.MockEthereumNode.WebSocket do
   # This module is heavily inspired by WebSockex.TestServer.
   # See https://github.com/Azolo/websockex/blob/master/test/support/test_server.ex
   @behaviour :cowboy_websocket
