+++ apps/potterhat_node/test/support/mock_listener.ex
@@ -1,4 +1,4 @@
-  defmodule Potterhat.Node.MockListener do
+  defmodule PotterhatNode.MockListener do
     use GenServer
 
     #
